add: endpoints for feedback, analyze and evaluate audio
