+++ API/evaluation_api/routers/evaluation_router.py
@@ -1,7 +1,11 @@
 from fastapi import APIRouter, UploadFile, Depends, HTTPException, Form
 from typing import Dict, Any
-from schemas.evaluation_schema import AnalysisResponse, FeedbackResponse
+
+from schemas.evaluation_schema import AnalysisResponse, EvaluationResponse, FeedbackResponse
 from services.analysis_service import AnalysisService, get_analysis_service
+from services.evaluation_service import EvaluationService, get_evaluation_service
+from services.feedback_service import FeedbackService, get_feedback_service
+from services.database_service import DatabaseService, get_database_service
 
 router = APIRouter(
     prefix="/evaluation",
@@ -15,7 +19,6 @@ async def root():
 @router.post("/analyze_audio", response_model=AnalysisResponse)
 async def analyze_audio(
     audio_file: UploadFile,
-    audio_id: str = Form(...),
     analysis_service: AnalysisService = Depends(get_analysis_service)
 ): 
     # Make sure an audio file was passed
@@ -26,9 +29,9 @@ async def analyze_audio(
         )
         
     try:
-        # Call the audio analysis function
-        analysis_results = await analysis_service.analyze_audio(audio_file, audio_id)
-        return analysis_results
+        # Call the audio analysis service
+        analysis_response = await analysis_service.analyze_audio(audio_file)
+        return analysis_response
     
     except Exception as e:
         print(f"Unexpected error: {str(e)}")
@@ -36,12 +39,15 @@ async def analyze_audio(
             status_code=500,
             detail="An unexpected error occurred during audio analysis."
         )
-        
-@router.post("/feedback", response_model=FeedbackResponse)
-async def analyze_audio(
+
+@router.post("/evaluate_audio", response_model=EvaluationResponse)
+async def evaluate_audio(
     audio_file: UploadFile,
     audio_id: str = Form(...),
-    analysis_service: AnalysisService = Depends(get_analysis_service)
+    resource_type:int = Form(...),
+    evaluation_service: EvaluationService = Depends(get_evaluation_service),
+    analysis_service: AnalysisService = Depends(get_analysis_service),
+    database_service:DatabaseService = Depends(get_database_service)
 ): 
     # Make sure an audio file was passed
     if not audio_file.content_type.startswith("audio/"):
@@ -51,13 +57,61 @@ async def analyze_audio(
         )
         
     try:
-        # Call the audio analysis function
-        analysis_results = await analysis_service.analyze_audio(audio_file, audio_id)
-        return analysis_results
+        # Get audio analysis
+        audio_analysis = await analysis_service.analyze_audio(audio_file)
+        
+        # Get reference analysis
+        reference_analysis = await database_service.get_reference_analysis(audio_id,
+                                                                           resource_type)
+        
+        # Call the audio evaluation service
+        evaluation_response = await evaluation_service.evaluate_audio(audio_analysis.model_dump(),
+                                                                      reference_analysis.model_dump())
+        
+        return evaluation_response
     
     except Exception as e:
         print(f"Unexpected error: {str(e)}")
         raise HTTPException(
             status_code=500,
-            detail="An unexpected error occurred during audio analysis."
+            detail="An unexpected error occurred during audio evaluation."
         )
+
+
+@router.post("/feedback", response_model=FeedbackResponse)
+async def feedback(
+    audio_file: UploadFile,
+    audio_id: str = Form(...),
+    resource_type:int = Form(...),
+    feedback_service: FeedbackService = Depends(get_feedback_service),
+    analysis_service: AnalysisService = Depends(get_analysis_service),
+    database_service:DatabaseService = Depends(get_database_service)
+): 
+    # Make sure an audio file was passed
+    if not audio_file.content_type.startswith("audio/"):
+        raise HTTPException(
+            status_code=400,
+            detail="Invalid file type. Please upload an audio file."
+        )
+        
+    try:
+        # Get audio analysis
+        audio_analysis = await analysis_service.analyze_audio(audio_file)
+        
+        # Get reference analysis
+        reference_analysis = await database_service.get_reference_analysis(audio_id,
+                                                                           resource_type)
+        
+        # Call the audio feedback function
+        feedback_response = await feedback_service.generate_feedback(audio_analysis.model_dump(),
+                                                                     reference_analysis.model_dump())
+        
+        return feedback_response
+    
+    except Exception as e:
+        print(f"Unexpected error: {str(e)}")
+        raise HTTPException(
+            status_code=500,
+            detail="An unexpected error occurred during audio feedback."
+        )
+
