+++ speech_evaluation/config.py.example
@@ -6,6 +6,6 @@ from dotenv import load_dotenv
 load_dotenv()
 
 # Access variables
-ARLI_API_KEY = os.getenv("API_KEY")
+ARLI_API_KEY = os.getenv("ARLI_API_KEY")
 if ARLI_API_KEY is None:
-    raise ValueError("API_KEY not found in environment variables.")
+    raise ValueError("ARLI_API_KEY not found in environment variables.")
